16-11-2018

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: aljacque <aljacque@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 16:06:38 by aljacque          #+#    #+#             */
-/*   Updated: 2018/11/07 16:33:12 by aljacque         ###   ########.fr       */
+/*   Updated: 2018/11/16 14:42:30 by aljacque         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int		i;
+	size_t				i;
 	unsigned const char	*cpy1;
 	unsigned const char	*cpy2;
