+++ ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: aljacque <aljacque@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/08 14:28:04 by aljacque          #+#    #+#             */
-/*   Updated: 2018/11/12 12:57:18 by aljacque         ###   ########.fr       */
+/*   Updated: 2018/11/16 14:48:51 by aljacque         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	if (n == 0)
 		return (0);
 	while ((unsigned char)s1[i] == (unsigned char)s2[i]
-	&& (unsigned int)i < n - 1)
+	&& i < n - 1)
 	{
 		if ((unsigned char)s1[i] == '\0')
 			return (0);
