+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: aljacque <aljacque@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/08 10:43:48 by aljacque          #+#    #+#             */
-/*   Updated: 2018/11/12 12:58:05 by aljacque         ###   ########.fr       */
+/*   Updated: 2018/11/16 14:49:11 by aljacque         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
-	unsigned int i;
+	size_t i;
 
 	i = 0;
 	while (src[i] != '\0' && i < n)
