+++ ft_strndup.c
@@ -6,7 +6,7 @@
 /*   By: aljacque <aljacque@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/14 10:00:58 by aljacque          #+#    #+#             */
-/*   Updated: 2018/11/14 10:07:55 by aljacque         ###   ########.fr       */
+/*   Updated: 2018/11/16 14:49:42 by aljacque         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 char	*ft_strndup(const char *str, size_t size)
 {
-	int		i;
+	size_t	i;
 	char	*dest;
 
 	i = 0;
 	if (!(dest = malloc(sizeof(char) * (size + 1))))
 		return (0);
-	while (str[i] != '\0' && (unsigned)i < size)
+	while (str[i] != '\0' && i < size)
 	{
 		dest[i] = str[i];
 		i++;
